Folder rename for components.
