+++ src/modules/purchase/components/PurchaseRoutingComponents.tsx
@@ -1,25 +1,20 @@
 import { Layout } from '@layout/components/base';
-import { PurchaseApprovalDetail } from '@purchase/components/purchaseHistories/detail/PurchaseApprovalDetail';
-import { PurchaseApprovalList } from '@purchase/components/purchaseHistories/list/PurchaseApprovalList';
-// import { PurchaseApprovalCollectionPage } fro@purchase/components/purchaseHistories/list/PurchaseApprovalListage';
-import { PurchaseRequestDetail } from '@purchase/components/purchaseRequest/detail/PurchaseRequestDetail';
-import { PurchaseRequestEditor } from '@purchase/components/purchaseRequest/editor/PurchaseRequestEditor';
-import { PurchaseRequestList } from '@purchase/components/purchaseRequest/list/PurchaseRequestList';
-// import { PurchaseRequestCollectionPage } fro@purchase/components/purchaseRequest/list/PurchaseRequestListage';
-import { PurchaseSettlementDetail } from '@purchase/components/purchaseSettlement/detail/PurchaseSettlementDetail';
-import { PurchaseSettlementEditor } from '@purchase/components/purchaseSettlement/editor/PurchaseSettlementEditor';
-import { PurchaseSettlementList } from '@purchase/components/purchaseSettlement/list/PurchaseSettlementList';
-// import { PurchaseSettlementCollectionPage } from '@purchase/components/purchaseSettlement/list/PurchaseSettlementList';
-import { SettlementApprovalDetail } from '@purchase/components/settlementHistories/detail/SettlementApprovalDetail';
-import { SettlementApprovalList } from '@purchase/components/settlementHistories/list/SettlementApprovalList';
-// import { SettlementApprovalCollectionPage } from '@purchase/components/settlementHistories/list/SettlementApprovalCollectionPage';
+import { PurchaseApprovalDetail } from '@purchase/components/request/purchaseApproval/detail/PurchaseApprovalDetail';
+import { PurchaseApprovalList } from '@purchase/components/request/purchaseApproval/list/PurchaseApprovalList';
+import { PurchaseRequestDetail } from '@purchase/components/request/purchaseRequest/detail/PurchaseRequestDetail';
+import { PurchaseRequestEditor } from '@purchase/components/request/purchaseRequest/editor/PurchaseRequestEditor';
+import { PurchaseRequestList } from '@purchase/components/request/purchaseRequest/list/PurchaseRequestList';
+import { PurchaseSettlementDetail } from '@purchase/components/settlement/purchaseSettlement/detail/PurchaseSettlementDetail';
+import { PurchaseSettlementEditor } from '@purchase/components/settlement/purchaseSettlement/editor/PurchaseSettlementEditor';
+import { PurchaseSettlementList } from '@purchase/components/settlement/purchaseSettlement/list/PurchaseSettlementList';
+import { SettlementApprovalDetail } from '@purchase/components/settlement/settlementApproval/detail/SettlementApprovalDetail';
+import { SettlementApprovalList } from '@purchase/components/settlement/settlementApproval/list/SettlementApprovalList';
 import * as React from 'react';
 import { Route, RouteComponentProps, Switch } from 'react-router';
 
 const purchase = (props: RouteComponentProps) => (
   <Switch>
     <Route path={`${props.match.path}/list`} component={PurchaseRequestList} />
-    {/* <Route path={`${props.match.path}/list`} component={PurchaseRequestCollectionPage} /> */}
     <Route path={`${props.match.path}/details/:purchaseUid`} component={PurchaseRequestDetail} />
     <Route path={`${props.match.path}/form`} component={PurchaseRequestEditor} />
   </Switch>
@@ -28,14 +23,12 @@ const purchase = (props: RouteComponentProps) => (
 const purchaseApproval = (props: RouteComponentProps) => (
   <Switch>
     <Route path={`${props.match.path}/list`} component={PurchaseApprovalList} />
-    {/* <Route path={`${props.match.path}/list`} component={PurchaseApprovalCollectionPage} /> */}
     <Route path={`${props.match.path}/details/:purchaseUid`} component={PurchaseApprovalDetail} />
   </Switch>
 );
 const settlement = (props: RouteComponentProps) => (
   <Switch>
     <Route path={`${props.match.path}/list`} component={PurchaseSettlementList} />
-    {/* <Route path={`${props.match.path}/list`} component={PurchaseSettlementCollectionPage} /> */}
     <Route path={`${props.match.path}/details/:purchaseUid`} component={PurchaseSettlementDetail} />    
     <Route path={`${props.match.path}/form`} component={PurchaseSettlementEditor} />
   </Switch>
@@ -44,7 +37,6 @@ const settlement = (props: RouteComponentProps) => (
 const settlementApproval = (props: RouteComponentProps) => (
   <Switch>
     <Route path={`${props.match.path}/list`} component={SettlementApprovalList} />
-    {/* <Route path={`${props.match.path}/list`} component={SettlementApprovalCollectionPage} /> */}
     <Route path={`${props.match.path}/details/:purchaseUid`} component={SettlementApprovalDetail} />
   </Switch>
 );
